FS-5062 - Transferring test of create_export_zip function (from blueprints/application/routes.py module) from general test_config_export into correct test module
